bin/emerge: Added removal of --tree and --changelog upon a restart of emerge. Added --ask support to emerge metadata.
pym/portage.py: Removed the requisite that a profile must depend on
">=sys-apps/portage-2.0.51" to be able to use package.mask.

bin/repoman pym/portage.py: Removed the disabling of PORTDIR_OVERLAY from
portage.py and added code to repoman to set PORTDIR to the current
directory's repo unless it doesn't contain a profiles directory. If that is
the case, PORTDIR is kept and PORTDIR_OVERLAY is set to the current overlay
-only-.