bin/emake: Made the make command configurable but defaulted to "make".
bin/emerge: Fixed an issue that would cause binary package reinstalls when
IUSE disagreed.

bin/repoman: Fixed an issue where the portage version wasn't appended to
the commit message when specified on the command line. Fixed the detection
of files above the package level during Manifest generation.

pym/portage.py: Renamed the userpriv FEATURE to userfetch for the fetch()
phase and fixed a bug where privileges would try to be dropped even as
non-root. Fixed a bug where files were attempted to be fetched when only a
manifest recreation is requested.