bin/ebuild.sh: Made /etc/profile.env sourced regardless of USERLAND. Made copying of config.sub and config.guess --force'd as they are sometimes u-w. Replaced a if/else of chgrp root/wheel with chgrp 0.
bin/emerge: Moved the system package unmerge check after slots are checked
and removed the countdown for when --ask is specified. Added a chdir to the
removal of old cache call so that there are no errors due to userpriv and
being in a non-portage-readable directory.

bin/etc-update: Moved the temporary work dir from /tmp/$$ to
${PORTAGE_TMPDIR}/$$.

bin/prepinfo bin/prepman: Reordered the -{min,max}depth and -name arguments
to find so that warnings are not produced.

pym/portage.py: Added support for FreeBSD's ldconfig. Fixed an incorrect
chown of BUILD_PREFIX instead of PORT_LOGDIR.